Fix option parsing skeleton
This fixes the option parsing structure so that it continues to operate
correctly. The `shift`s are left in the 'after' and 'before' cases
because at this point it is highly likely they will be used to get the
required date parameters for each option. It is also likely that option
parsing will need to be done a bit differently in order to match
provided dates or to account for the '--after=<date>' form.