Make install follow convention
This makes the Makefile follow normal configure/make/make install
convention, using [1] as a reference. For this purpose a 'configure'
script is also added. It's a pretty simple install flow so it seems a
bit silly to include all this but it will make it easier if one wants to
use some sort of packaging software for it.

The 'configure' script will generate a 'Makefile.in' if options are
given (currently only install prefix). The Makefile will also obey the
'DESTDIR' variable although it is not included in the 'configure'
script.

[1]: http://nuclear.mutantstargoat.com/articles/make/#going-the-extra-mile-for-release