Update option syntax to match git-log
This makes the after/since before/until syntax match that of `git log`
and also adds `sed` commands to set variables for the given options.

All that is left to be done now is update how parameters are given to
the '_summary' function and this can be merged to master. Perhaps [1]
will help.

[1]: https://stackoverflow.com/questions/3236871/how-to-return-a-string-value-from-a-bash-function