Remove functions to avoid needing to pass options
The _parseArgs function needed to pass the given arguments to the
_summary function but since bash only supports status return codes this
would have been difficult. Plus the function structure was unnecessary
anyway since the whole thing executed in order and the function blocks
were not repeated.

This removes the function structure and adds the options into the block
which prints the summary information, meaning git-summary now
understands the after/since before/until options of git log.