Fix option values with whitespace bug
As a result of how quoting in bash works values provided to the
after/since and before/until options were stripped of their quotes. This
means doing something like `git summary --since="yesterday midnight"`
would fail, while `git log --since="yesterday midnight"` is a perfectly
acceptable command.

This changes the 'opts' variable to be an array in order to preserve
quoting for use in the 'git log' commands.

This also adds checks that will assign the 'adds' and 'subs' variables
to zero when they are empty avoiding error messages from the 'numfmt'
command.