Add color printing
This adds color printing using git to get whether or not we should print
in color (so that this command will be uniform with the rest of git in
handling colors). It also uses git to get the color values so we aren't
directly using ANSI escape codes and so that colors are configurable.

This sets it up so that user additions print in color.diff.new
(defaulting to green) and user deletions print in color.diff.old
(defaulting to red). More colors could be added later, but for now these
two alone seem reasonable.